+++ users/RELEASES.md
@@ -29,3 +29,13 @@ Based on:
 - [typescript v2.0.1] users
 ### Releases
 - [NPM v2.0.1] https://www.npmjs.com/package/@wingspan/users/v/2.0.1 - users
+
+## 2023-11-27 01:30:11
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470eea6a7577e003be20a26
+- Speakeasy CLI 1.122.0 (2.195.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.0.2] users
+### Releases
+- [NPM v2.0.2] https://www.npmjs.com/package/@wingspan/users/v/2.0.2 - users